addItem function

+++ 14-Grocery-List/app.js
@@ -27,19 +27,72 @@ window.addEventListener("DOMContentLoaded", setupItems);
 // ****** FUNCTIONS **********
 
 function addItem(e){
+    e.preventDefault();
+    const value = grocery.value;
+    const id = new Date().getTime().toString();
+
+    if (value && !editFlag){
+        //console.log("submiting")
+
+        const element = document.createElement('article');
+
+        // add id
+        let attr = document.createAttribute('data-id');
+        attr.value = id;
+        element.setAttributeNode(attr);
+        
+        // add class
+        element.classList.add('grocery-item');
+
+        // change innerHtml
+        element.innerHTML = `<p class="title">${value}</p>
+            <div class="btn-container">
+              <!-- edit btn -->
+              <button type="button" class="edit-btn">
+                <i class="fas fa-edit"></i>
+              </button>
+              <!-- delete btn -->
+              <button type="button" class="delete-btn">
+                <i class="fas fa-trash"></i>
+              </button>
+            </div>
+          `;
+
+        // event listeners for both buttons on new items
+        const deleteBtn = element.querySelector(".delete-btn");
+        deleteBtn.addEventListener("click", deleteItem);
+        const editBtn = element.querySelector(".edit-btn");
+        editBtn.addEventListener("click", editItem);
 
 
-    // even listeners for both buttons on new items
-    const deleteBtn = element.querySelector(".delete-btn");
-    deleteBtn.addEventListener("click", deleteItem);
-    const editBtn = element.querySelector(".edit-btn");
-    editBtn.addEventListener("click", editItem);
+        displayAlert("added item to list", "success");
+        container.classList.add("show-container");
+
+        // almost done
+
+    }
+    else if (value && editFlag){
+        //console.log("editing")
+
+        editElement.innerHTML = value;
+        displayAlert("value changed", "success");
+
+    }
+    else{
+        displayAlert("Please enter value", "danger");
+    }
 
 }
 
 
 function displayAlert(text, action){
+    alert.textContent = text;
+    alert.classList.add(`alert-${action}`);
 
+    setTimeout(() => {
+        alert.textContent = "";
+        alert.classList.remove(`alert-${action}`);
+    }, 1000);
 }
 
 function clearItems(){
